Changed the way night time is rendered, worked on fire animations, and some audio.

+++ MyGame-core/src/physics/Lighting/Fire.java
@@ -21,7 +21,7 @@ public class Fire extends GameObject {
 
 	private boolean hasLogs;
 
-	public Fire(float x, float y, int width, int height, String location, boolean hasLogs) {
+	public Fire(float x, float y, float width, float height, String location, boolean hasLogs) {
 		this.x                = x;
 		this.y                = y;
 		this.width            = width;
